switch to using inventory holders instead of name checking
