+++ paper/run/server.properties
@@ -1,5 +1,5 @@
 #Minecraft server properties
-#Sun Nov 05 15:18:16 EST 2023
+#Sun Nov 05 16:19:23 EST 2023
 enable-jmx-monitoring=false
 rcon.port=25575
 level-seed=
