+++ paper/src/main/java/us/crazycrew/crazycrates/paper/CrazyHandler.java
@@ -12,7 +12,7 @@ import us.crazycrew.crazycrates.paper.api.crates.PrizeManager;
 import us.crazycrew.crazycrates.paper.api.support.metrics.MetricsWrapper;
 import us.crazycrew.crazycrates.paper.api.users.BukkitUserManager;
 import us.crazycrew.crazycrates.paper.api.support.structures.blocks.ChestManager;
-import us.crazycrew.crazycrates.paper.api.users.guis.InventoryManager;
+import us.crazycrew.crazycrates.paper.api.crates.menus.InventoryManager;
 import us.crazycrew.crazycrates.paper.commands.CommandManager;
 import us.crazycrew.crazycrates.paper.modules.ModuleLoader;
